NO-TICKET Add infrastructure for multiple loggers
This commit adds the ability for the system to support multiple logger instances, while still
including the default Log utility that exists. This change will allow users of the SDK to
implement their own loggers and handle log messages as they see fit. As well, this mirrors the
capabilities of the iOS SDK.

This change is achieved by introducing an interface for those custom loggers (FRLogger),
implementing that interface with the Android Log class (FRConsoleLogger), and adding a set of
logger instances to the Logger class that can be added or removed (addLogger, removeLogger).